refactor utils, delete last_used field and other
